Refactor: Componentize modal from ItemCard component
